requirements.txt
requests
python-dotenv
pinecone-client
langchain
langchain-google-vertexai
fastapi
uvicorn
psycopg2-binary
python-dotenv
pinecone-client
google-generativeai
jinja2
pandas
openpyxl
sqlalchemy 
psycopg2-binary
flask
matplotlib